+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Fix laptop names not being able to be updated
+
 ## 0.0.1 / 2012-11-20
 
 * Initial revision