+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Only catch StandardError, not Exception, in Turntabler#run
 * Fix song votes being tracked with empty user ids
 * Fix the current dj's points not getting updated on song_voted events
 * Fix song_ended event never getting fired