+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Authenticate using emails / passwords instead of user ids / auth tokens
 * Fix exceptions in callbacks not causing one-time callbacks to be unregistered
 * Only catch StandardError, not Exception, in Turntabler#run
 * Fix song votes being tracked with empty user ids